+++ clang/lib/Sema/SemaExpr.cpp
@@ -18011,6 +18011,7 @@ HandleImmediateInvocations(Sema &SemaRef,
                            Sema::ExpressionEvaluationContextRecord &Rec) {
   if ((Rec.ImmediateInvocationCandidates.size() == 0 &&
        Rec.ReferenceToConsteval.size() == 0) ||
+      Rec.isImmediateFunctionContext() ||
       SemaRef.RebuildingImmediateInvocation)
     return;